+++ Changelog.md
@@ -1,4 +1,5 @@
 # Master
+  * New core / mapping API (devteamreims/4ME#163)
   * Update webpack to latest version + tweak build script
   * Autocomplete background color is now black to provide better contrast (devteamreims/4ME#167)
   * ControlRoom widget will now display a tooltip on hover instead of change the widget title (devteamreims/4ME#164)